% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_2_ddct.R
\name{calculate_2_ddct}
\alias{calculate_2_ddct}
\title{calculate_2_ddct}
\usage{
calculate_2_ddct(data, ref_gene, control_name)
}
\arguments{
\item{data}{A table input with special sequence columns
(from left 2 right as follow:sample,treatment,biological_replicate,
refer gene name,your gene name1,your gene name2,..),
each column split with tab and containing header.}

\item{ref_gene}{your refer gene name}

\item{control_name}{the control condition name in the treatment column
of data.}
}
\value{
the result of 2(-delta delta CT).
}
\description{
The function is used for calculating the
2(-delta delta CT) and the result is stored in a column
named containing "_2_ddct".
}
\examples{
# example
data <- data.frame(
  sample = rep(rep(1:4, each = 3), 2),
  treatment = rep(c("Control", "Treated"), each = 12),
  your_refer_gene = c(
    19.68, 19.69, 19.80, 19.95, 19.93, 19.97, 19.93, 20.02, 20.27,
    19.93, 19.88, 19.90, 20.61, 19.98, 20.57, 19.68, 19.95,
    19.85, 20.27, 20.08, 20.07, 20.10, 20.07, 20.10
  ),
  gene01 = c(
    23.22, 23.34, 23.12, 24.06, 24.15, 24.15, 23.18, 23.13,
    23.10, 24.78, 24.45, 24.67, 23.11, 22.99, 23.10, 22.77,
    22.99, 23.06, 23.73, 24.01, 23.80, 23.73, 23.83, 23.73
  ),
  gene02 = c(
    29.08, 29.04, 29.39, 28.23, 28.01, 28.12, 28.79, 28.43,
    28.49, 31.37, 30.74, 31.09, 27.11, 27.24, 27.37, 25.52,
    25.72, 25.52, 27.43, 26.73, 26.65, 27.96, 27.84, 27.98
  )
)

# Using calculate_dd_ct calculating 2(-delta delta CT)
dd.ct <- calculate_2_ddct(data, "your_refer_gene", "Control")
head(dd.ct)
# Values in gene01_2_ddct and gene02__2_ddct are 2(-delta delta CT)
}
