% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ddct_t_tests.R
\name{ddct_t_tests}
\alias{ddct_t_tests}
\title{ddct_t_tests}
\usage{
ddct_t_tests(data, dd.ct, control)
}
\arguments{
\item{data}{A table input with special sequence columns
(from left 2 right as follow:sample,treatment,biological_replicate,
refer gene name,your gene name1,your gene name2,..),
each column split with tab and containing header.}

\item{dd.ct}{the output of function of Calculate_dd_ct.}

\item{control}{the control condition name}
}
\value{
t_test_results
}
\description{
t test is used to perform Statistical Analysis of −ΔΔCt
between control and each treatment.
}
\examples{
# Create an example data frame
dd.ct <- data.frame(
  biological_replicate = c(1, 1, 2, 2, 3, 3, 4, 4),
  treatment = c("Control", "Treated", "Control", "Treated", "Control", "Treated", "Control", "Treated"),
  gene01 = c(3.50, 2.68, 4.17, 3.11, 3.06, 3.71, 4.73, 3.67),
  gene02 = c(9.45, 6.85, 8.17, 5.76, 8.50, 6.80, 11.2, 7.84),
  treatment_control_vals = rep("Control", 8),
  gene01_control_vals = rep(3.50, 8),
  gene02_control_vals = rep(9.45, 8),
  gene01_diff = c(0, -0.82, 0, -1.06, 0, 0.65, 0, -1.06),
  gene01_2_ddct = c(1, 0.66, 1, 0.68, 1, 1.38, 1, 0.68),
  gene02_diff = c(0, -2.60, 0, -2.41, 0, -1.65, 0, -3.36),
  gene02_2_ddct = c(1, 0.73, 1, 0.70, 1, 0.81, 1, 0.70)
)

# Print the example data frame
print(dd.ct)
# example
data <- data.frame(
  sample = rep(rep(1:4, each = 3), 2),
  treatment = rep(c("Control", "Treated"), each = 12),
  your_refer_gene = c(
    19.68, 19.69, 19.80, 19.95, 19.93, 19.97, 19.93, 20.02, 20.27,
    19.93, 19.88, 19.90, 20.61, 19.98, 20.57, 19.68, 19.95,
    19.85, 20.27, 20.08, 20.07, 20.10, 20.07, 20.10
  ),
  gene01 = c(
    23.22, 23.34, 23.12, 24.06, 24.15, 24.15, 23.18, 23.13,
    23.10, 24.78, 24.45, 24.67, 23.11, 22.99, 23.10, 22.77,
    22.99, 23.06, 23.73, 24.01, 23.80, 23.73, 23.83, 23.73
  ),
  gene02 = c(
    29.08, 29.04, 29.39, 28.23, 28.01, 28.12, 28.79, 28.43,
    28.49, 31.37, 30.74, 31.09, 27.11, 27.24, 27.37, 25.52,
    25.72, 25.52, 27.43, 26.73, 26.65, 27.96, 27.84, 27.98
  )
)
control <- "Control"
# t-test
t_test_results <- ddct_t_tests(data, dd.ct, control = "Control")
}
