% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qpcr_plot.R
\name{qpcr_plot}
\alias{qpcr_plot}
\title{qpcr_plot}
\usage{
qpcr_plot(
  dd.ct = dd.ct,
  gene = gene,
  control = control,
  order = NULL,
  t_test_results = NULL,
  y_lab = NULL,
  bar_size = 1,
  bar_width = 0.8,
  errorbar_width = 0.2,
  errorbar_size = 1,
  bar_position = 0.8,
  bar_colors = NULL,
  p_val_labelsize = 6,
  axis_text_x_size = 14,
  axis_text_y_size = 14,
  axis_title_y_size = 18,
  y_max = NULL,
  panel.border.size = 1,
  axis.ticks.size = 1,
  rotation_angle = 0,
  axis.text.x.dis = 0.5,
  axis.ticks.length = 0.15
)
}
\arguments{
\item{dd.ct}{the output of function of Calculate_dd_ct.}

\item{gene}{the qPCR result of gene you want to show.}

\item{t_test_results}{the output of function of perform_t_tests}

\item{bar_width}{modify the width of the bars}

\item{bar_position}{modify the spacing between the bars}

\item{bar_colors}{modify the colors of the bars}

\item{p_val_labelsize}{modify the size of p-value}

\item{axis_text_x_size}{modify the size of the x-axis text}

\item{axis_text_y_size}{modify the size of the y-axis text}

\item{axis_title_y_size}{modify the size of the y-axis title}

\item{errorbar_bar}{modify the width of the errorbar}

\item{plot_width}{modify the width of plot}

\item{plot.height}{modify the height of plot}
}
\value{
Bar plot of qPCR result for single gene in different treatments.
}
\description{
Visualize the qPCR result through bar plot
for single gene between Control and single/different treatments.
}
\examples{
dd.ct <- data.frame(
  biological_replicate = c(1, 1, 2, 2, 3, 3, 4, 4),
  treatment = c("Control", "Treated1", "Control", "Treated1", "Control", "Treated1", "Control", "Treated1"),
  gene01 = c(3.50, 2.68, 4.17, 3.11, 3.06, 3.71, 4.73, 3.67),
  gene02 = c(9.45, 6.85, 8.17, 5.76, 8.50, 6.80, 11.2, 7.84),
  treatment_control_vals = rep("Control", 8),
  gene01_control_vals = rep(3.50, 8),
  gene02_control_vals = rep(9.45, 8),
  gene01_diff = c(0, -0.82, 0, -1.06, 0, 0.65, 0, -1.06),
  gene01_2_ddct = c(1, 0.66, 1, 0.68, 1, 1.38, 1, 0.68),
  gene02_diff = c(0, -2.60, 0, -2.41, 0, -1.65, 0, -3.36),
  gene02_2_ddct = c(1, 0.73, 1, 0.70, 1, 0.81, 1, 0.70)
)
t_test_results <- data.frame(
  column_name = c("gene01_2_ddct", "gene02_2_ddct"),
  group1 = c("Control", "Control"),
  group2 = c("Treated1", "Treated1"),
  p_value = c(0.255771789, 0.004903698),
  significance = c("ns", "**")
)
gene <- "gene01"
control <- "Control"
qpcr_plot(
  dd.ct = dd.ct, gene = gene, control = control, order = NULL, t_test_results = t_test_results,
  y_lab = NULL, bar_size = 1, bar_width = 0.8, errorbar_width = 0.2, errorbar_size = 1, bar_position = 0.8, bar_colors = NULL, p_val_labelsize = 6,
  axis_text_x_size = 14, axis_text_y_size = 14, axis_title_y_size = 18,
  y_max = NULL, panel.border.size = 1, axis.ticks.size = 1, rotation_angle = 0, axis.text.x.dis = 0.5, axis.ticks.length = 0.15
)
}
